pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "ronnyx"
version = "0.2.0"
description = "Self-hosted MCP runtime with custom tool support. Connect any MCP server, add your own tools, chat via API or CLI."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Baran Yalcin" }
]

dependencies = [
  "fastapi>=0.111.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "langchain-core>=1.0.0,<2.0.0",
  "langchain-openai>=1.0.0,<2.0.0",
  "langgraph>=1.0.0,<2.0.0",
  "langchain-mcp-adapters>=0.2.0,<1.0.0",
  "requests>=2.31.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[project.scripts]
ronnyx-chat = "ronnyx.cli:main"
ronnyx-serve = "ronnyx.serve:main"

[project.urls]
Homepage = "https://github.com/baranylcn/ronnyx"
Source = "https://github.com/baranylcn/ronnyx"
Issues = "https://github.com/baranylcn/ronnyx/issues"

[tool.hatch.build.targets.wheel]
packages = ["ronnyx"]

[tool.hatch.build.targets.sdist]
include = ["ronnyx", "README.md", "pyproject.toml", "ronnyx.yaml.example"]
